Update copy of egress firewall objects
Egress firewall objects have an assigned status which needs to be
updated by OVN-Kubernetes. The update can however lead to a data race
with the API server object in case the update is not done on a copy.
This is something that has been seen when testing PR:
https://github.com/ovn-org/ovn-kubernetes/pull/1640 which enables the
race detector

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>