make egressfirewall use logical_router_policies
Change egressfirewall to use logical_router_policies on the
ovn_cluster_router as oppsed to using ACLs on each nodes join
switches.

This changes optimizes the generation of egressfirewall policies
because we only have to create one object and put it in one place
the ovn_cluster_router applies to all namespaces so we only need
to attach the policy there.

because the single router applies to all namespaces the test
"correctly adds an existing egressFirewall to a new node" is no longer needed

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>