Remove node management of egress IP, no longer needed!
We don't need this anymore in "new local gateway" mode. In fact we can
regard this new mode as "shared" gateway mode for what concerns egress IP.
We need the master to setup logical routing policies to the egress node and then
use OVN NAT rules to NAT to the egress IP. This is because traffic exists
a node through ovn-k8s-mp0 in this new mode and allows the NAT-ing to be done
on OVN's side without additional iptable rules on the host.
Since we also have the conntrack flows programmed on the external bridge, we
are also insured that traffic flowing back to OVN gets back to the egress pods.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>