Fixes requiring host-network for exgw pods in update
If an ovn network pod was updated with exgw annotations, we were
ignoring it because it was not rescheduled and was not a host networked
pod. We should always be updating the exgw for an ovn networked pod
because the pod may be using multus annotations and not rely on being
host networked.

Signed-off-by: Tim Rozet <trozet@redhat.com>