This commit adds support for auditing the raft membership for NB and SB databases as well as a health-check to recover from a corrupted cluster.
Following use-cases are supposed to be solved:

- When ovnkube-db pods are restarted in a raft (HA)
mode, they will have a new server id with which they will join the cluster.
This will cause the members list in the db to have two entries for the same node.

- When any of the existing master nodes is replaced, the cluster will end up
showing the old master's sid in the members list. The replaced node needs to
be evicted. This commit adds support to detect an orphan raft member and kick
it out of the cluster.

- Finally, when any one local instance of the raft db is corrupted,
today there is no mechanism for the db to recover itself automatically.
With this commit, we add a mechanism (via the ovn-dbchecker container) to monitor
individual raft dbs for cluster-status and in case the db is reported as inconsistent/corrupt,
we backup the db files on that node and restart the corresponding ovsdb-server
container. This will force the db to get initialized.

This does not handle the case when the entire raft db gets corrupted. That will need a manual intervention akin to performing the following steps:

1. Stop ovsdb-server (nb/sb) instances on all nodes.
2. Backup and delete all the db files while the server instances are down.
3. Bring up the ovsdb-server instances to start with a fresh raft cluster id
4. Reset ovn-northd instances in case of nbdb corruption by calling nb-cluster-state-reset
5. Reset ovn-controller instances in case of sbdb corruption by calling sb-cluster-state-reset
6. Restart the source of truth (ovn-k-master) to force sync the dbs in case nbdb got corrupted.
7. If it's purely sbdb corruption, then the ovn-northd will propagate the changes to the sbdb,
    so there is no need to restart the ovn-k-master in that case.

Co-authored-by: Tim Rozet <trozet@redhat.com>
Signed-off-by: Aniket Bhat <anbhat@redhat.com>