informer: Skip update revision check for tests
The informer package won't process updates under test conditions
when the informers have been started. This is due to a check
to ensure the objects K8s revision number has incremented and
therefore this is genuine work that we want to enqueue. That
revision number isn't implemented in the fakeClient.
Rather than implement this, which would be a lot of work, I
introduced the EventHandlerFactory type, and allowed a
function that matches this signature to be passed in to the
NewNode and NewMaster functions.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>