Fix ovn_cluster_router policy deletion
When we delete a node we initiate gateway cleanup which deletes PBR on
ovn_cluster_router. In the current code we list all PBR rules and then
use string submatch to figure out what to delete. Inter node policy was
not including the node name, so it was accidentally deleting PBR rules
for other nodes.

Additionally, the match logic was flawed in the case where there are
node names that overlap. For example, ovn-worker and ovn-worker2. If
ovn-worker node was deleted, it may accidentally match and delete
policies for ovn-worker2.

Both of these issues are addressed by this patch.

Signed-off-by: Tim Rozet <trozet@redhat.com>