enable e2e multicast testing
Create a multicast source in one node and 2 listener in another
node.
One of the listener should join the group and receive the multicast
traffic, and the other listener should try to join another group
that does not exist and fail to receive any multicast traffic.

Organize the e2e folder correctly to avoid issues with go modules and
imports.

only run ha and local gw for control-plane e2e

Signed-off-by: Antonio Ojea <aojea@redhat.com>