Consider changed configuration when kicking out raft members
In certain use-cases such as master replacement, we have seen
that the configuration of the db/master pods that holds the addresses
of the raft members isn't updated by the time the new master joins the
raft cluster. When this happens the db pods carrying the old address
configuration for nbdb and sbdb would kick the newly joined master out.
This commit addresses this issue by additionally checking existing master
pods' IPs against the server id list for determining whether a given
server id should be booted out of the raft cluster.

Signed-off-by: Aniket Bhat <anbhat@redhat.com>