Migrate default deny multicast policy to port-group
Currently default deny multicast policy creates
a global port-group, add all the logical-ports
across the cluster and associate an multicast deny
acl to that port-group. This acl, also uses the
port-group as a inport and outport match.

To apply ACL to the logical_switch, port-group
just need one port to be part of the port-group.
Adding all the logical-port of a switch to the
port-group is unnecessary.

Given that the deny acl will be applied across
the switch, installing acl that matches on all
the port is also unnecessary.

This commit simplifies the acl by removing the match
on inport/outport, and associate it to the port group
that has only management ports of all the logical-switches
of the cluster nodes.

Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>
Co-authored-by: Dan Williams <dcbw@redhat.com>