Fix egress node reconciliation
PR: https://github.com/ovn-org/ovn-kubernetes/pull/1668 re-worked
the way egress IP assignment is done, taking into account that users
do not accidentally assign a node IP as egress IP and screws up cluster
networking. However, it introduced a bug, namely: we now have a problem
with reconciliation of nodes for egress, ultimately skipping any UPDATE
event for nodes. Thus, if ovnkube-node does not label a node in question with
k8s.ovn.org/node-primary-ifaddr before ovnkube-master received the
ADD for that node, then we never proceed to update our internal
data with that node when we receive the UPDATE. This commit fixes that

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>