add a periodic job to run against k8s master
We had a periodic job to test dual-stack, this job was added with
the intention of getting signal on dual-stack, so we could observe
the results.

After the job was stable, it was promoted to presubmit, so we can
gate on dual-stack.

New dual-stack features are being added upstream, so we should test
against those. In order to do that we modify current dual-stack job
to use master kubernetes.

Signed-off-by: Antonio Ojea <aojea@redhat.com>