Fix: update of egress IP for empty re-assignments
Unrelated to the previous commits, this one fixes an update issue
during the re-assignment procedure, namely: if a re-assignment is
attempted when there are no more nodes to assign to, we previously
exited the function `reassignEgressIP` without updating the status.
This commit thus changes the execution to continue on to the update
phase, yet log the error once the re-assignment is finished.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>