move const decleration for the ovn_cluster_router to a new package
create a types package to and move constants from util into it.
Beginning to refactor the ovn package and split the resources into there
own packages

In addition make the constants closer in line with the golang style conventions
for initlaisms

https://github.com/golang/go/wiki/CodeReviewComments#initialisms

"Words in names that are initialisms or acronyms (e.g. "URL" or "NATO")
have a consistent case."

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>