move const decleration for the ovn_cluster_router to the config package
Moving the const declerations to a more suitable location. Moving from the
util package to the config package. I believe this is more in line with
the style of other golang projects.

In addition make the constants closer in line with the golang style conventions
for initlaisms

https://github.com/golang/go/wiki/CodeReviewComments#initialisms

"Words in names that are initialisms or acronyms (e.g. "URL" or "NATO")
have a consistent case."

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>