Remove duplicate steps code with composite actions
With two e2e job sections now (one for PRs and pushes) some of the
duplicated steps can be referenced from a composite actions.yml
file to reduce the duplication and help with maintenance.

github composite run actions will add more support for things
like 'if' in the future and we will be able to further de-dup.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>