egress_ip: Move functions to egressIPController
This commit formalizes egressIPController as the place where all
egressIP related functions live.

It moves a number of egressIP related functions out of the OvnController
as they only require access to the WatchFactory - I'll be working on
factoring this out in future.

When moving these functions over I noticed that EgressIP was fetching from
the API Server and not from the cache. Once I'd fixed this to use the
informer cache there was only UpdateEgressIP left in the Kube interface.
As it's only used in the egress IP code, I went ahead and in-lined it in
egress_ip.go to save passing the Kube interface around everywhere.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>