Enable debug logging for unit test runs
Some unit test debugging could be made easier by enabling
debug log level. This PR changes the log level in PrepareTestConfig
which is used from the tests, as opposed to the LoggingConfig object.
This was done as to not impact users of OVN-Kubernetes deploying a
cluster without specifying --loglevel and modifying the default setting

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>