test: Set GOMAXPROCS to 1
GOMAXPROCS will default to whatever is offered by the host CPU.
As we are running in a container, we should respect whatever
CPU limits are imposed on the container. While we could use
automaxprocs to do this, it is simpler for our test cases
to set GOMAXPROCS to 1.

I hope this will solve a number of timing issues that have recently
plagued our unit tests.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>