Add timeout to kind delete
Deleting the ovn-kubernetes namespace can leave the command hanging,
because there is a loss of connectivity to the cluster while tearing
down all of hte networking pods. This patch makes the command timeout
after 5 seconds, which when added to the sleep of 5 seconds afterwards
should be enough time for the pods to be torn down to avoid the Fedora
31/32 crash that happens if the pods are not wiped before removing the
kind docker containers.

Signed-off-by: Tim Rozet <trozet@redhat.com>