Fixes ICMPv6 flooding behavior
Our flows for flooding icmpv6 to both OVN and the host were too generic.
This caused ICMPv6 echo requests to be sent to both, resulting in a
duplicate ack and other issues. This patch changes the flows for icmpv6
to be more specific and target only flooding for Neighbor Advertisements
and Route Advertisements. This will allow OVN to receive only the
packets it needs for peering/discovery, while all other ICMPv6 will go to
the host only.

Signed-off-by: Tim Rozet <trozet@redhat.com>