expose Kubernetes clientset in the OVN controller
The OVN controller was embedding the Kubernetes clientset
under an Interface, thus other methods couldn't use it.

We add a new field in the controller with the Clientset,
so we don't have to modify the kube.Interface each time
that we need a new functionality, that the Kubernetes
clientset provide us for free.

Signed-off-by: Antonio Ojea <aojea@redhat.com>