fixing few raft db startup script issues
the change addressing the following errors:

1. on init startup, pod-1 and pod-2 suppose to be waiting on a critical section
while pod-0 is initializing, however, it isn't working as expected as
a result pod-0 entering infinite CLB

2. restart on leader causing election timer set to fail with error
"election timer change pending"

3. restart on non-leader node, we may still see code path entering into
initial create section

4. replacement of a node following restart of one of existing 2 nodes would
enter into inifinte CLB

5. if the last OVN DB pod fails before it creates the OVN DB service endpoint,
it restarts without trying again to create the endpoint

with the change we introduce retry on rapid timer set when the previous is in
pending state and remove assumption that pod-0 is always a leader

we got rid of $initial_raft_create as we already checking for leader in
more reliable way. This variable is also not properly reset, which
makes logic difficult to understand.

Signed-off-by: Dmitry Yusupov <dyusupov@nvidia.com>
Co-authored-by: Yun Zhou <yunz@nvidia.com>
Acked-by: Girish Moodalbail <gmoodalbail@nvidia.com>