run services repair controller only at startup
The repair loop in the new Services controller checks the services
present in the cache and compares them with the services present
in the OVN database, deleting the OVN load balancers that are not
associated to a service.

This can be racy, also, it's unlikely that this happen meanwhile the
controller is running, but it is necessary for upgrades or situations
that can cause that both OVN and Kubenetes get out of sync.

We run the repair once at startup, just after the informer caches are
syncer, so we start with a clean state.

Signed-off-by: Antonio Ojea <aojea@redhat.com>