don't compile k8s in the CI
We are compiling the e2e.test and kubect binaries from the
kubernetes repo to match the Kubernetes version used in kind.

However, this binaries doesn't need to match the exact version
of the cluster under test, and we can use precompiled binaries
that are publicly available, saving more than 500Mb of storage and
5 minutes in the CI.

It's also a great benefit for local developments, so you don't
need to checkout and recompile the kubernetes sources.

Signed-off-by: Antonio Ojea <aojea@redhat.com>