informer: EventHandler now handles updates
The initial implementation of this package did not cater for updates.
This was intentional as add functions should be idempotent.
However, as this pattern is not implemented throughout the rest of
the codebase and changing those functions could take a very long time,
it seems sensible to instead adjust this package.

For this to work, we encode the eventType as part of the key in the
workqueue. This is less efficient, because we can't collapse all
event types for a given item so only the most recent is processed.
However, it is more efficient than processing every event as we
do today.

Additionally, we must also cache old copies of an object such that
we can provide the update event handler with the old/new objects.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>