support multipath to detect default gateway
We were detecting the default gateway based in the assumption that
the netlink library was populating the de Gw field with it, however,
it turns out that when multipath is enabled that field is empty and
the corresponding information is in the MultiPath field.

Also, in order to obtain the default gateway, we were dumping all
routes filtering by FAMILY_ALL, and iterating on the results to get
the ones we are interested. We can filter directly by the IP Family instead,
to avoid future possibles problems with the FAMILY_ALL interpretion.

Signed-off-by: Antonio Ojea <aojea@redhat.com>