SDN-1492: Add ovnkube-trace binary to image
The network-tools repo requires ovnkube-trace for the scripts. Would
be nice if the openshift/ovn-kubernetes ships the binary so that
the binary can copied to the network-tools image. The alternative is
to vendor the whole ovn-kubernetes repo which cause pulling a lot
of dependencies. Considering the image already builds the dbchecker,
adding ovnkube-trace as well would only make it more consistent.