Move egress firewall to ACLs in local GW mode
Currently the features egress firewall and egress IP are mutually
exclusive. This is due to the fact that both use logical router policies
to determine routing decisions, hence only one or the either can take
precedence given the fact that they work agnostically of one another.

Currently egress firewall applies logical router policies with the
highest priority, egress IP does so with the lowest. This means that if
a pod matches on both a egress firewall and egress IP, the packet exiting
the cluster will not have the egress IP.

This should however be the case if something like this is specified by the user:

```
apiVersion: k8s.ovn.org/v1
kind: EgressFirewall
metadata:
  name: default
  namespace: default
spec:
  egress:
  - to:
      cidrSelector: 172.18.0.5/32
    type: Allow
  - to:
      cidrSelector: 0.0.0.0/0
    type: Deny
---
apiVersion: k8s.ovn.org/v1
kind: EgressIP
metadata:
  name: egressip-dup
spec:
  egressIPs:
  - 172.18.0.45
  podSelector:
    matchLabels:
      name: netserver-0
  namespaceSelector:
    matchLabels:
      name: default
```

To be able to have these features work at the same time we'll need to
move egress firewall from any routing rules and instead apply ACLs, which is its
"natural construct" for blocking / allowing traffic.

This only needs to be done in local GW mode, since in shared GW mode this is
already the case. We need to apply the ACLs though to all node switches to have it
work with regular network policies (which apply their ACLs to the local switch port).

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>